refactor uuid
