+++ weldx/asdf/tags/weldx/__init__.py
@@ -1,1 +1,12 @@
-from . import aws, core, debug, equipment, groove, measurement, processes, time, unit
+from . import (
+    aws,
+    base_types,
+    core,
+    debug,
+    equipment,
+    groove,
+    measurement,
+    processes,
+    time,
+    unit,
+)
