+++ weldx/asdf/tags/weldx/core/common_types.py
@@ -1,6 +1,5 @@
 from dataclasses import dataclass
 from typing import List
-from uuid import UUID
 
 import numpy as np
 import pint
@@ -9,26 +8,6 @@ from pandas.api.types import is_timedelta64_dtype as is_timedelta
 
 from weldx.asdf.types import WeldxType
 from weldx.constants import WELDX_QUANTITY as Q_
-
-
-# UUID ---------------------------------------------------------------------------------
-class UuidTypeASDF(WeldxType):
-    """Implements a version 4 UUID."""
-
-    name = "uuid"
-    version = "1.0.0"
-    types = [UUID]
-    requires = ["weldx"]
-
-    @classmethod
-    def to_tree(cls, node: UUID, ctx):
-        """convert to python dict"""
-        return dict(uuid=str(node))
-
-    @classmethod
-    def from_tree(cls, tree, ctx):
-        """Reconstruct form tree."""
-        return UUID(tree["uuid"])
 
 
 # Dimension ----------------------------------------------------------------------------
